Moved decorator notimplemented to Decorators.
